differentSwapTypes
PRICING A 5-YEAR SEMI-ANNUAL AMORTIZING INTEREST RATE SWAPS

The pricing of an interest rate swap involves finding the fixed-rate  for the swap. 
Conventionally, the notional of an IRS stays constant over the life of the swap but 
how about if the notional does not stay the same but either accretes or
amortizes as we move towards maturity? These types of swaps are known as
accreting and amortizing swaps respectively. In order to price an interest
rate swap, it is necessary to perform the following steps:
    
    
    Step 1: Construct a cash curve
    
    Step 2: Derive a zero coupon curve
    
    Step 3: Calculate the discount factors to find present values
    
    Step 4: Derive an implied forward curve
    
    Step 5: Find the present value of the floating side of the swap.
    
    Step 6: Use numerical methods or optimization to find the fixed swap rate.
    
    Step 7: Equate the P.V. of the floating side, with the P.V of the fixed side
    and this find the value of the fixed rate.