Create CDF_blackScholes
"""
This is an implementation to calculate Cummulative Density Function (CDF) and Option Price (in Black-Scholes Model) 
using the Pseudocode from the Primer of Mathematics for Financial Engineering by Dan Stefanica.  I will check to see how 
the results compare to the ones generated by Matlab's Financial Instrument ToolBox at a later time.  But for now this is just an 
implementation of the code from the above book.  Black-Scholes is an analytical formula and good for valuing instruments
that cannot be exercised before maturity such as European Call or Put.  To value options such as American Put or call
that can be exercised prior to maturity, you may have to use binomial models or trees.  Check out Financial Engineering
and Risk Management on Coursera, the course gave a good treatment of these topics (Thank goodness for these free courses, 
you can just learn on a continous basis (as time permits) and don't have to pay a dime).  I will post codes for building 
binomial lattices later (but it may take a long while).

COMPARISON OF CODE VERSUS MATLAB FINANCIAL INSTRUMENT TOOLBOX
You can use Matlab to price options as well by calling like so:
[Call, Put] = blsprice(Price, Strike, Rate, Time, Volatility, Yield)
So if you replace the parameters above with the following values you get the following prices which converges to the 
ones generated by this python module:
[Call, Put] = blsprice(42, 40, 0.05, 0.5, 0.3, 0.03)

Call = 4.7053

Put = 2.3430


If you run this code in python, like so
(blackScholes(0,42.0,40.0,0.5,0.3,0.05,0.03,'call'))
(blackScholes(0,42.0,40.0,0.5,0.3,0.05,0.03,'put'))

You get same results as above:
The value of a call with the given parameters is : 4.7053

The value of a put with the given parameters is : 2.3430



"""